split into separate files
made tokenizer and singlewordtokenizer their own files. added debugging logs to naivebayesclassifier for testing.
