+++ src/ITokenizer.ts
@@ -1,0 +1,7 @@
+
+interface Tokenizer {
+    tokenize(a:string): string[];
+
+}
+
+export default Tokenizer;
