+++ src/SingleWordTokenizer.ts
@@ -1,0 +1,10 @@
+import Tokenizer from "./ITokenizer";
+
+class SingleWordTokenizer implements Tokenizer {
+    tokenize(a:string): string[] {
+        return a.toLocaleLowerCase().replace(/[.,\/#!$%\^&\*;:{}=\-_`~()]/g,"").split("");
+        
+    }
+}
+
+export default SingleWordTokenizer;
